Split utils.js tests to utils/selection.js and utils/structure.js.
